Fix #10444: Client Window Factory
