+++ docs/14_0_0/core/clientwindow.md
@@ -6,6 +6,20 @@ This allows for separating state not only per session but also on a per client-w
 
 
 Prime Client Window is activated in faces-config.xml
+
+**MyFaces:**
+```xml
+<faces-config>
+    ...
+    <factory>
+        <client-window-factory>org.primefaces.clientwindow.PrimeClientWindowFactory</client-window-factory>
+    </factory>
+</faces-config>
+```
+
+**Mojarra:** see: https://github.com/eclipse-ee4j/mojarra/issues/5297
+
+```xml
 ```xml
 <faces-config>
     ...
@@ -15,8 +29,9 @@ Prime Client Window is activated in faces-config.xml
 </faces-config>
 ```
 
-It´s a replacement for the standard JSF-mechanism activated in web.xml. So only add PrimeClientWindowLifecycleFactory
-but not add javax.faces.CLIENT_WINDOW_MODE!
+It´s a replacement for the standard JSF-mechanism activated in web.xml. So only add `PrimeClientWindowLifecycleFactory` or `PrimeClientWindowFactory`
+but do **NOT** add `javax.faces.CLIENT_WINDOW_MODE`!
+
 ```xml
 <context-param>
     <param-name>javax.faces.CLIENT_WINDOW_MODE</param-name>
@@ -29,8 +44,8 @@ See [MultiViewState](core/multiviewstate.md) for details.
 
 Prime Client Window adds jfwid-parameter to all url´s rendered by JSF and PrimeFaces.
 Prime Client Window offers improved handling compared to JSF´s built in mechanism in following points:
-* JSF doesnt append the url param on initial redirect, so f5 creates a new windowId
-* JSF doesnt validate if the tab was initialy openend with another windowId
+* JSF does NOT append the url param on initial redirect, so f5 creates a new windowId
+* JSF does NOT validate if the tab was initially opened with another windowId
 * it uses sessionStorage to save and validate the windowId
 * the workflow is the same as Apache DeltaSpike lazy mode
 
@@ -42,4 +57,4 @@ Prime Client Window offers improved handling compared to JSF´s built in mechani
 
 
 ## Further Notes
-JSF 4.0 also may add ClientWindowScoped OOTB - see https://github.com/eclipse-ee4j/faces-api/issues/1509
+JSF 4.0 has added ClientWindowScoped OOTB - see https://github.com/eclipse-ee4j/faces-api/issues/1509
